+++ app/src/main/java/app/passwordstore/injection/prefs/PreferenceModule.kt
@@ -41,4 +41,9 @@ class PreferenceModule {
   fun providePGPPassphrases(@ApplicationContext context: Context): SharedPreferences {
     return context.getSharedPreferences("${BuildConfig.APPLICATION_ID}_passphrases", MODE_PRIVATE)
   }
+
+  @[Provides GitSecrets Reusable]
+  fun provideGitSecrets(@ApplicationContext context: Context): SharedPreferences {
+    return context.getSharedPreferences("${BuildConfig.APPLICATION_ID}_git_secrets", MODE_PRIVATE)
+  }
 }