+++ app/src/main/java/app/passwordstore/util/settings/PreferenceKeys.kt
@@ -17,7 +17,6 @@ object PreferenceKeys {
   )
   const val CLEAR_CLIPBOARD_20X = "clear_clipboard_20x"
   const val CLEAR_CLIPBOARD_HISTORY = "clear_clipboard_history"
-  @Deprecated(message = "We have retired persistent caching of passwords")
   const val CLEAR_SAVED_PASS = "clear_saved_pass"
   const val COPY_ON_DECRYPT = "copy_on_decrypt"
   const val ENABLE_DEBUG_LOGGING = "enable_debug_logging"
@@ -54,7 +53,6 @@ object PreferenceKeys {
 
   @Deprecated("Git branch is no longer stored in preferences")
   const val GIT_BRANCH_NAME = "git_branch"
-  @Deprecated(message = "Now unused, as we have retired persistent caching of the https password")
   const val HTTPS_PASSWORD = "https_password"
   const val LENGTH = "length"
   const val OREO_AUTOFILL_CUSTOM_PUBLIC_SUFFIXES = "oreo_autofill_custom_public_suffixes"
@@ -82,9 +80,6 @@ object PreferenceKeys {
   const val SSH_KEY = "ssh_key"
   const val SSH_KEYGEN = "ssh_keygen"
 
-  @Deprecated(
-    message = "Now unused, as we have retired persistent caching of the ssh key passphrase"
-  )
   const val SSH_KEY_LOCAL_PASSPHRASE = "ssh_key_local_passphrase"
 
   const val SSH_OPENKEYSTORE_CLEAR_KEY_ID = "ssh_openkeystore_clear_keyid"